module9-task3. Добавлен компонент MainEmpty. Добавлен условный рендер: когда нет предложений, то на главной странице отрисовывается MainEmpty.
