+++ src/components/main-empty/main-empty.tsx
@@ -1,0 +1,22 @@
+type MainEmptyProps = {
+  currentCityName: string;
+}
+
+function MainEmpty({currentCityName}: MainEmptyProps): JSX.Element {
+  return (
+    <>
+      <section className="cities__no-places">
+        <div className="cities__status-wrapper tabs__content">
+          <b className="cities__status">No places to stay available</b>
+          <p className="cities__status-description">We could not find any property available at the moment in {currentCityName}</p>
+        </div>
+      </section>
+      <div className="cities__right-section">
+        <div className="cities__map-container" />
+      </div>
+    </>
+  );
+}
+
+export default MainEmpty;
+
