--- src/pages/main-empty/main-empty.tsx
@@ -1,49 +1,0 @@
-import {Link} from 'react-router-dom';
-import CitiesList from '../../components/cities-list/cities-list';
-import {City} from '../../types/city';
-import HeaderUserProfile from '../../components/header-user-profile/header-user-profile';
-
-type MainEmptyProps = {
-    cities: City[];
-}
-
-function MainEmpty({cities}: MainEmptyProps): JSX.Element {
-  return (
-    <div className="page page--gray page--main">
-      <header className="header">
-        <div className="container">
-          <div className="header__wrapper">
-            <div className="header__left">
-              <Link className="header__logo-link header__logo-link--active" to="#">
-                <img className="header__logo" src="img/logo.svg" alt="6 cities logo" width="81" height="41" />
-              </Link>
-            </div>
-            <HeaderUserProfile />
-          </div>
-        </div>
-      </header>
-
-      <main className="page__main page__main--index page__main--index-empty">
-        <h1 className="visually-hidden">Cities</h1>
-        <div className="tabs">
-          <section className="locations container">
-            <CitiesList cities={cities} />
-          </section>
-        </div>
-        <div className="cities">
-          <div className="cities__places-container cities__places-container--empty container">
-            <section className="cities__no-places">
-              <div className="cities__status-wrapper tabs__content">
-                <b className="cities__status">No places to stay available</b>
-                <p className="cities__status-description">We could not find any property available at the moment in Dusseldorf</p>
-              </div>
-            </section>
-            <div className="cities__right-section"></div>
-          </div>
-        </div>
-      </main>
-    </div>
-  );
-}
-
-export default MainEmpty;
